Display a list of all the available output objects

Function generating a comprehensive list of available output variables (series, parameters, diagnostics) from the estimation process by the x13(), regarima() and x11() functions. Some items are available in the default estimation output but the remainder can be added using the userdefined parameter. User-defined objects can the be retrieved from the list of lists generated by the estimation process

Usage

userdefined_variables_x13(x = c("X-13", "RegArima", "X-11"))

Arguments

x

a character to indicate the estimation function for which the output items list will be displayed.

Value

a vector containing the names of all the available output objects (series, diagnostics, parameters)

References

More information and examples related to 'JDemetra+' features in the online documentation: https://jdemetra-new-documentation.netlify.app/

Examples

userdefined_variables_x13("x13")
#>   [1] "adjust"                              
#>   [2] "arima.bd"                            
#>   [3] "arima.bp"                            
#>   [4] "arima.bphi(*)"                       
#>   [5] "arima.bq"                            
#>   [6] "arima.btheta(*)"                     
#>   [7] "arima.d"                             
#>   [8] "arima.p"                             
#>   [9] "arima.phi(*)"                        
#>  [10] "arima.q"                             
#>  [11] "arima.theta(*)"                      
#>  [12] "benchmarking.original"               
#>  [13] "benchmarking.result"                 
#>  [14] "benchmarking.target"                 
#>  [15] "cal"                                 
#>  [16] "cal_b"                               
#>  [17] "cal_b(?)"                            
#>  [18] "cal_f"                               
#>  [19] "cal_f(?)"                            
#>  [20] "decomposition.b1"                    
#>  [21] "decomposition.b10"                   
#>  [22] "decomposition.b11"                   
#>  [23] "decomposition.b13"                   
#>  [24] "decomposition.b17"                   
#>  [25] "decomposition.b2"                    
#>  [26] "decomposition.b20"                   
#>  [27] "decomposition.b3"                    
#>  [28] "decomposition.b4"                    
#>  [29] "decomposition.b5"                    
#>  [30] "decomposition.b6"                    
#>  [31] "decomposition.b7"                    
#>  [32] "decomposition.b8"                    
#>  [33] "decomposition.b9"                    
#>  [34] "decomposition.c1"                    
#>  [35] "decomposition.c10"                   
#>  [36] "decomposition.c11"                   
#>  [37] "decomposition.c13"                   
#>  [38] "decomposition.c17"                   
#>  [39] "decomposition.c2"                    
#>  [40] "decomposition.c20"                   
#>  [41] "decomposition.c4"                    
#>  [42] "decomposition.c5"                    
#>  [43] "decomposition.c6"                    
#>  [44] "decomposition.c7"                    
#>  [45] "decomposition.c9"                    
#>  [46] "decomposition.d1"                    
#>  [47] "decomposition.d10"                   
#>  [48] "decomposition.d11"                   
#>  [49] "decomposition.d12"                   
#>  [50] "decomposition.d13"                   
#>  [51] "decomposition.d2"                    
#>  [52] "decomposition.d4"                    
#>  [53] "decomposition.d5"                    
#>  [54] "decomposition.d6"                    
#>  [55] "decomposition.d7"                    
#>  [56] "decomposition.d8"                    
#>  [57] "decomposition.d9"                    
#>  [58] "decomposition.d9-global-msr"         
#>  [59] "decomposition.d9-msr"                
#>  [60] "decomposition.d9-msr-table"          
#>  [61] "decomposition.i_cmp"                 
#>  [62] "decomposition.s_cmp"                 
#>  [63] "decomposition.s_cmp_b"               
#>  [64] "decomposition.s_cmp_f"               
#>  [65] "decomposition.sa_cmp"                
#>  [66] "decomposition.sa_cmp_b"              
#>  [67] "decomposition.sa_cmp_f"              
#>  [68] "decomposition.seasonal-filters"      
#>  [69] "decomposition.si_cmp"                
#>  [70] "decomposition.t_cmp"                 
#>  [71] "decomposition.t_cmp_b"               
#>  [72] "decomposition.t_cmp_f"               
#>  [73] "decomposition.trend-filter"          
#>  [74] "decomposition.x11-all"               
#>  [75] "decomposition.y_cmp"                 
#>  [76] "decomposition.y_cmp_b"               
#>  [77] "decomposition.y_cmp_f"               
#>  [78] "det"                                 
#>  [79] "det_b(?)"                            
#>  [80] "det_f(?)"                            
#>  [81] "det_i"                               
#>  [82] "det_i_b(?)"                          
#>  [83] "det_i_f(?)"                          
#>  [84] "det_s"                               
#>  [85] "det_s_b(?)"                          
#>  [86] "det_s_f(?)"                          
#>  [87] "det_t"                               
#>  [88] "det_t_b(?)"                          
#>  [89] "det_t_f(?)"                          
#>  [90] "diagnostics.fcast-insample-mean"     
#>  [91] "diagnostics.fcast-outsample-mean"    
#>  [92] "diagnostics.fcast-outsample-variance"
#>  [93] "diagnostics.seas-i-combined"         
#>  [94] "diagnostics.seas-i-combined3"        
#>  [95] "diagnostics.seas-i-evolutive"        
#>  [96] "diagnostics.seas-i-f"                
#>  [97] "diagnostics.seas-i-friedman"         
#>  [98] "diagnostics.seas-i-kw"               
#>  [99] "diagnostics.seas-i-periodogram"      
#> [100] "diagnostics.seas-i-qs"               
#> [101] "diagnostics.seas-i-spectralpeaks"    
#> [102] "diagnostics.seas-i-stable"           
#> [103] "diagnostics.seas-lin-combined"       
#> [104] "diagnostics.seas-lin-evolutive"      
#> [105] "diagnostics.seas-lin-f"              
#> [106] "diagnostics.seas-lin-friedman"       
#> [107] "diagnostics.seas-lin-kw"             
#> [108] "diagnostics.seas-lin-periodogram"    
#> [109] "diagnostics.seas-lin-qs"             
#> [110] "diagnostics.seas-lin-spectralpeaks"  
#> [111] "diagnostics.seas-lin-stable"         
#> [112] "diagnostics.seas-res-combined"       
#> [113] "diagnostics.seas-res-combined3"      
#> [114] "diagnostics.seas-res-evolutive"      
#> [115] "diagnostics.seas-res-f"              
#> [116] "diagnostics.seas-res-friedman"       
#> [117] "diagnostics.seas-res-kw"             
#> [118] "diagnostics.seas-res-periodogram"    
#> [119] "diagnostics.seas-res-qs"             
#> [120] "diagnostics.seas-res-spectralpeaks"  
#> [121] "diagnostics.seas-res-stable"         
#> [122] "diagnostics.seas-sa-ac1"             
#> [123] "diagnostics.seas-sa-combined"        
#> [124] "diagnostics.seas-sa-combined3"       
#> [125] "diagnostics.seas-sa-evolutive"       
#> [126] "diagnostics.seas-sa-f"               
#> [127] "diagnostics.seas-sa-friedman"        
#> [128] "diagnostics.seas-sa-kw"              
#> [129] "diagnostics.seas-sa-periodogram"     
#> [130] "diagnostics.seas-sa-qs"              
#> [131] "diagnostics.seas-sa-spectralpeaks"   
#> [132] "diagnostics.seas-sa-stable"          
#> [133] "diagnostics.seas-si-combined"        
#> [134] "diagnostics.seas-si-combined3"       
#> [135] "diagnostics.seas-si-evolutive"       
#> [136] "diagnostics.seas-si-stable"          
#> [137] "diagnostics.td-i-all"                
#> [138] "diagnostics.td-i-last"               
#> [139] "diagnostics.td-res-all"              
#> [140] "diagnostics.td-res-last"             
#> [141] "diagnostics.td-sa-all"               
#> [142] "diagnostics.td-sa-last"              
#> [143] "ee"                                  
#> [144] "ee_b(?)"                             
#> [145] "ee_f(?)"                             
#> [146] "finals.d11"                          
#> [147] "finals.d11a"                         
#> [148] "finals.d11b"                         
#> [149] "finals.d12"                          
#> [150] "finals.d12a"                         
#> [151] "finals.d12b"                         
#> [152] "finals.d13"                          
#> [153] "finals.d16"                          
#> [154] "finals.d16a"                         
#> [155] "finals.d16b"                         
#> [156] "finals.d18"                          
#> [157] "finals.d18a"                         
#> [158] "finals.d18b"                         
#> [159] "finals.e1"                           
#> [160] "finals.e11"                          
#> [161] "finals.e2"                           
#> [162] "finals.e3"                           
#> [163] "i"                                   
#> [164] "l"                                   
#> [165] "likelihood.adjustedll"               
#> [166] "likelihood.aic"                      
#> [167] "likelihood.aicc"                     
#> [168] "likelihood.bic"                      
#> [169] "likelihood.bic2"                     
#> [170] "likelihood.bicc"                     
#> [171] "likelihood.df"                       
#> [172] "likelihood.hannanquinn"              
#> [173] "likelihood.ll"                       
#> [174] "likelihood.neffectiveobs"            
#> [175] "likelihood.nobs"                     
#> [176] "likelihood.nparams"                  
#> [177] "likelihood.ssqerr"                   
#> [178] "log"                                 
#> [179] "m-statistics.m1"                     
#> [180] "m-statistics.m10"                    
#> [181] "m-statistics.m11"                    
#> [182] "m-statistics.m2"                     
#> [183] "m-statistics.m3"                     
#> [184] "m-statistics.m4"                     
#> [185] "m-statistics.m5"                     
#> [186] "m-statistics.m6"                     
#> [187] "m-statistics.m7"                     
#> [188] "m-statistics.m8"                     
#> [189] "m-statistics.m9"                     
#> [190] "m-statistics.q"                      
#> [191] "m-statistics.q-m2"                   
#> [192] "mhe"                                 
#> [193] "mhe_b(?)"                            
#> [194] "mhe_f(?)"                            
#> [195] "omhe"                                
#> [196] "omhe_b(?)"                           
#> [197] "omhe_f(?)"                           
#> [198] "out"                                 
#> [199] "out_b(?)"                            
#> [200] "out_f(?)"                            
#> [201] "out_i"                               
#> [202] "out_i_b(?)"                          
#> [203] "out_i_f(?)"                          
#> [204] "out_s"                               
#> [205] "out_s_b(?)"                          
#> [206] "out_s_f(?)"                          
#> [207] "out_t"                               
#> [208] "out_t_b(?)"                          
#> [209] "out_t_f(?)"                          
#> [210] "period"                              
#> [211] "preadjustment.a1"                    
#> [212] "preadjustment.a1a"                   
#> [213] "preadjustment.a1b"                   
#> [214] "preadjustment.a6"                    
#> [215] "preadjustment.a7"                    
#> [216] "preadjustment.a8"                    
#> [217] "preadjustment.a8i"                   
#> [218] "preadjustment.a8s"                   
#> [219] "preadjustment.a8t"                   
#> [220] "preadjustment.a9"                    
#> [221] "preadjustment.a9sa"                  
#> [222] "preadjustment.a9ser"                 
#> [223] "preadjustment.a9u"                   
#> [224] "reg_i"                               
#> [225] "reg_i_b(?)"                          
#> [226] "reg_i_f(?)"                          
#> [227] "reg_s"                               
#> [228] "reg_s_b(?)"                          
#> [229] "reg_s_f(?)"                          
#> [230] "reg_sa"                              
#> [231] "reg_sa_b(?)"                         
#> [232] "reg_sa_f(?)"                         
#> [233] "reg_t"                               
#> [234] "reg_t_b(?)"                          
#> [235] "reg_t_f(?)"                          
#> [236] "reg_u"                               
#> [237] "reg_u_b(?)"                          
#> [238] "reg_u_f(?)"                          
#> [239] "reg_y"                               
#> [240] "reg_y_b(?)"                          
#> [241] "reg_y_f(?)"                          
#> [242] "regression.description"              
#> [243] "regression.details.coefficients"     
#> [244] "regression.details.covar"            
#> [245] "regression.details.covar-ml"         
#> [246] "regression.easter"                   
#> [247] "regression.espan.end"                
#> [248] "regression.espan.missing"            
#> [249] "regression.espan.n"                  
#> [250] "regression.espan.start"              
#> [251] "regression.leaster"                  
#> [252] "regression.lp"                       
#> [253] "regression.mean"                     
#> [254] "regression.missing(*)"               
#> [255] "regression.ml.parameters"            
#> [256] "regression.ml.pcorr"                 
#> [257] "regression.ml.pcovar"                
#> [258] "regression.ml.pcovar-ml"             
#> [259] "regression.ml.pscore"                
#> [260] "regression.mu"                       
#> [261] "regression.nao"                      
#> [262] "regression.nlp"                      
#> [263] "regression.nls"                      
#> [264] "regression.nmh"                      
#> [265] "regression.nout"                     
#> [266] "regression.nso"                      
#> [267] "regression.ntc"                      
#> [268] "regression.ntd"                      
#> [269] "regression.nusers"                   
#> [270] "regression.out(*)"                   
#> [271] "regression.outlier(*)"               
#> [272] "regression.td(*)"                    
#> [273] "regression.type"                     
#> [274] "regression.user(*)"                  
#> [275] "residuals.bp"                        
#> [276] "residuals.bp2"                       
#> [277] "residuals.doornikhansen"             
#> [278] "residuals.kurtosis"                  
#> [279] "residuals.lb"                        
#> [280] "residuals.lb2"                       
#> [281] "residuals.lruns"                     
#> [282] "residuals.ludruns"                   
#> [283] "residuals.mean"                      
#> [284] "residuals.nruns"                     
#> [285] "residuals.nudruns"                   
#> [286] "residuals.res"                       
#> [287] "residuals.seasbp"                    
#> [288] "residuals.seaslb"                    
#> [289] "residuals.ser"                       
#> [290] "residuals.skewness"                  
#> [291] "residuals.tsres"                     
#> [292] "residuals.type"                      
#> [293] "s"                                   
#> [294] "s_b"                                 
#> [295] "s_f"                                 
#> [296] "sa"                                  
#> [297] "sa_b"                                
#> [298] "sa_f"                                
#> [299] "span.end"                            
#> [300] "span.missing"                        
#> [301] "span.n"                              
#> [302] "span.start"                          
#> [303] "t"                                   
#> [304] "t_b"                                 
#> [305] "t_f"                                 
#> [306] "tde"                                 
#> [307] "tde_b(?)"                            
#> [308] "tde_f(?)"                            
#> [309] "variancedecomposition.cycle"         
#> [310] "variancedecomposition.irregular"     
#> [311] "variancedecomposition.others"        
#> [312] "variancedecomposition.seasonality"   
#> [313] "variancedecomposition.tdh"           
#> [314] "variancedecomposition.total"         
#> [315] "y"                                   
#> [316] "y_b"                                 
#> [317] "y_b(?)"                              
#> [318] "y_eb(?)"                             
#> [319] "y_ef(?)"                             
#> [320] "y_f"                                 
#> [321] "y_f(?)"                              
#> [322] "yc"                                  
#> [323] "ycal"                                
#> [324] "ycal_f(?)"                           
userdefined_variables_x13("regarima")
#>   [1] "adjust"                          "arima.bd"                       
#>   [3] "arima.bp"                        "arima.bphi(*)"                  
#>   [5] "arima.bq"                        "arima.btheta(*)"                
#>   [7] "arima.d"                         "arima.p"                        
#>   [9] "arima.phi(*)"                    "arima.q"                        
#>  [11] "arima.theta(*)"                  "cal"                            
#>  [13] "cal_b(?)"                        "cal_f(?)"                       
#>  [15] "det"                             "det_b(?)"                       
#>  [17] "det_f(?)"                        "det_i"                          
#>  [19] "det_i_b(?)"                      "det_i_f(?)"                     
#>  [21] "det_s"                           "det_s_b(?)"                     
#>  [23] "det_s_f(?)"                      "det_t"                          
#>  [25] "det_t_b(?)"                      "det_t_f(?)"                     
#>  [27] "ee"                              "ee_b(?)"                        
#>  [29] "ee_f(?)"                         "l"                              
#>  [31] "likelihood.adjustedll"           "likelihood.aic"                 
#>  [33] "likelihood.aicc"                 "likelihood.bic"                 
#>  [35] "likelihood.bic2"                 "likelihood.bicc"                
#>  [37] "likelihood.df"                   "likelihood.hannanquinn"         
#>  [39] "likelihood.ll"                   "likelihood.neffectiveobs"       
#>  [41] "likelihood.nobs"                 "likelihood.nparams"             
#>  [43] "likelihood.ssqerr"               "log"                            
#>  [45] "mhe"                             "mhe_b(?)"                       
#>  [47] "mhe_f(?)"                        "omhe"                           
#>  [49] "omhe_b(?)"                       "omhe_f(?)"                      
#>  [51] "out"                             "out_b(?)"                       
#>  [53] "out_f(?)"                        "out_i"                          
#>  [55] "out_i_b(?)"                      "out_i_f(?)"                     
#>  [57] "out_s"                           "out_s_b(?)"                     
#>  [59] "out_s_f(?)"                      "out_t"                          
#>  [61] "out_t_b(?)"                      "out_t_f(?)"                     
#>  [63] "period"                          "reg_i"                          
#>  [65] "reg_i_b(?)"                      "reg_i_f(?)"                     
#>  [67] "reg_s"                           "reg_s_b(?)"                     
#>  [69] "reg_s_f(?)"                      "reg_sa"                         
#>  [71] "reg_sa_b(?)"                     "reg_sa_f(?)"                    
#>  [73] "reg_t"                           "reg_t_b(?)"                     
#>  [75] "reg_t_f(?)"                      "reg_u"                          
#>  [77] "reg_u_b(?)"                      "reg_u_f(?)"                     
#>  [79] "reg_y"                           "reg_y_b(?)"                     
#>  [81] "reg_y_f(?)"                      "regression.description"         
#>  [83] "regression.details.coefficients" "regression.details.covar"       
#>  [85] "regression.details.covar-ml"     "regression.easter"              
#>  [87] "regression.espan.end"            "regression.espan.missing"       
#>  [89] "regression.espan.n"              "regression.espan.start"         
#>  [91] "regression.leaster"              "regression.lp"                  
#>  [93] "regression.mean"                 "regression.missing(*)"          
#>  [95] "regression.ml.parameters"        "regression.ml.pcorr"            
#>  [97] "regression.ml.pcovar"            "regression.ml.pcovar-ml"        
#>  [99] "regression.ml.pscore"            "regression.mu"                  
#> [101] "regression.nao"                  "regression.nlp"                 
#> [103] "regression.nls"                  "regression.nmh"                 
#> [105] "regression.nout"                 "regression.nso"                 
#> [107] "regression.ntc"                  "regression.ntd"                 
#> [109] "regression.nusers"               "regression.out(*)"              
#> [111] "regression.outlier(*)"           "regression.td(*)"               
#> [113] "regression.type"                 "regression.user(*)"             
#> [115] "residuals.bp"                    "residuals.bp2"                  
#> [117] "residuals.doornikhansen"         "residuals.kurtosis"             
#> [119] "residuals.lb"                    "residuals.lb2"                  
#> [121] "residuals.lruns"                 "residuals.ludruns"              
#> [123] "residuals.mean"                  "residuals.nruns"                
#> [125] "residuals.nudruns"               "residuals.res"                  
#> [127] "residuals.seasbp"                "residuals.seaslb"               
#> [129] "residuals.ser"                   "residuals.skewness"             
#> [131] "residuals.tsres"                 "residuals.type"                 
#> [133] "span.end"                        "span.missing"                   
#> [135] "span.n"                          "span.start"                     
#> [137] "tde"                             "tde_b(?)"                       
#> [139] "tde_f(?)"                        "y"                              
#> [141] "y_b(?)"                          "y_eb(?)"                        
#> [143] "y_ef(?)"                         "y_f(?)"                         
#> [145] "yc"                              "ycal"                           
#> [147] "ycal_f(?)"                      
userdefined_variables_x13("x11")
#>  [1] "b1"               "b10"              "b11"              "b13"             
#>  [5] "b17"              "b2"               "b20"              "b3"              
#>  [9] "b4"               "b5"               "b6"               "b7"              
#> [13] "b8"               "b9"               "c1"               "c10"             
#> [17] "c11"              "c13"              "c17"              "c2"              
#> [21] "c20"              "c4"               "c5"               "c6"              
#> [25] "c7"               "c9"               "d1"               "d10"             
#> [29] "d11"              "d12"              "d13"              "d2"              
#> [33] "d4"               "d5"               "d6"               "d7"              
#> [37] "d8"               "d9"               "d9-global-msr"    "d9-msr"          
#> [41] "d9-msr-table"     "seasonal-filters" "trend-filter"     "x11-all"