Outlier Detection with a RegARIMA Model

Outlier Detection with a RegARIMA Model

Usage

regarima_outliers(
  y,
  order = c(0L, 1L, 1L),
  seasonal = c(0L, 1L, 1L),
  mean = FALSE,
  X = NULL,
  X.td = NULL,
  ao = TRUE,
  ls = TRUE,
  tc = FALSE,
  so = FALSE,
  cv = 0,
  clean = FALSE
)

Arguments

y

the dependent variable (a ts object).

order, seasonal

the orders of the ARIMA model.

mean

Boolean to include or not the mean.

X

user defined regressors (other than calendar).

X.td

calendar regressors.

ao, ls, so, tc

Boolean to indicate which type of outliers should be detected.

cv

numeric. The entered critical value for the outlier detection procedure. If equal to 0 the critical value for the outlier detection procedure is automatically determined by the number of observations.

clean

Clean missing values at the beginning/end of the series. Regression variables are automatically resized, if need be.

Value

a "JD3_REGARIMA_OUTLIERS" object, containing input variables and results

Examples

regarima_outliers(rjd3toolkit::ABS$X0.2.09.10.M)
#> $model
#> $model$y
#>         Jan    Feb    Mar    Apr    May    Jun    Jul    Aug    Sep    Oct
#> 1982                       460.1  502.6  443.8  459.1  438.4  465.1  452.7
#> 1983  379.2  378.0  472.1  503.4  510.6  462.4  468.3  458.2  482.7  485.3
#> 1984  414.7  414.5  484.7  487.3  597.9  500.4  543.4  503.4  522.8  556.6
#> 1985  516.3  452.5  525.8  587.7  700.3  561.8  602.8  582.5  563.1  637.1
#> 1986  570.5  478.2  547.4  594.3  751.6  553.4  663.2  581.1  661.9  665.6
#> 1987  613.9  513.2  599.9  674.1  714.0  670.5  720.9  601.6  672.3  709.1
#> 1988  631.0  551.1  678.1  715.7  740.8  722.0  683.5  650.9  723.3  729.6
#> 1989  631.5  552.0  719.0  697.6  764.8  786.3  715.1  723.8  757.9  751.7
#> 1990  678.2  586.2  726.8  744.1  815.5  832.4  710.3  759.4  741.1  786.6
#> 1991  694.0  604.7  719.2  748.2  828.2  746.9  794.5  770.4  741.5  858.6
#> 1992  740.0  665.9  701.5  831.4  878.6  826.0  788.2  723.6  819.8  902.5
#> 1993  762.1  643.0  754.1  840.7  906.6  887.1  771.5  728.7  844.7  886.9
#> 1994  745.7  664.4  821.5  831.7  908.0  912.6  782.9  798.8  887.0  934.6
#> 1995  752.4  682.5  811.2  906.0  927.2  906.8  880.6  873.9  856.8  920.6
#> 1996  833.1  737.1  812.0  895.2  962.8  908.6  908.0  888.9  833.7  933.7
#> 1997  840.9  727.4  857.9  849.0  994.8  830.2  971.1  836.0  939.1  976.9
#> 1998  917.3  716.2  822.9  970.1  970.2  849.4 1042.3  869.9  939.4 1021.3
#> 1999  942.0  738.4  903.2  953.2 1011.2  894.4 1054.5  899.5 1002.3 1043.7
#> 2000  924.9  798.2  901.9 1024.7 1052.3 1165.5  859.3 1009.2 1054.6 1070.4
#> 2001  971.9  814.6 1017.5 1039.2 1123.5 1024.9 1100.8  963.0 1012.9 1132.0
#> 2002 1027.9  841.4 1043.9 1075.3 1190.9 1143.0 1075.7 1065.9 1060.1 1211.4
#> 2003 1099.3  900.5 1092.7 1222.4 1237.1 1237.9 1182.0 1101.2 1198.2 1316.1
#> 2004 1182.9  989.8 1131.4 1277.1 1280.3 1384.1 1305.9 1166.8 1317.9 1358.3
#> 2005 1246.3 1037.3 1300.8 1153.7 1264.2 1454.2 1290.1 1210.7 1277.8 1314.4
#> 2006 1193.7 1037.7 1204.5 1348.6 1267.6 1429.0 1412.0 1239.2 1219.1 1344.6
#> 2007 1267.3 1047.0 1331.6 1302.6 1365.1 1491.5 1462.3 1315.5 1353.3 1440.6
#> 2008 1397.8 1140.5 1351.7 1396.6 1421.1 1401.6 1582.3 1268.4 1383.3 1452.4
#> 2009 1451.0 1056.6 1386.9 1509.1 1519.4 1500.5 1570.7 1341.5 1399.9 1534.3
#> 2010 1469.1 1111.9 1379.9 1389.7 1427.2 1551.4 1581.0 1324.0 1422.0 1464.9
#> 2011 1412.6 1117.5 1321.6 1472.6 1408.9 1471.9 1532.5 1293.5 1345.7 1404.7
#> 2012 1362.4 1131.7 1349.2 1391.2 1456.9 1616.4 1423.4 1359.0 1367.8 1442.6
#> 2013 1397.4 1113.6 1397.3 1339.1 1441.9 1537.4 1390.6 1337.2 1359.4 1463.3
#> 2014 1451.0 1064.9 1293.2 1442.9 1411.8 1461.6 1501.6 1254.2 1356.4 1478.7
#> 2015 1471.2 1053.8 1367.2 1442.2 1428.7 1480.9 1540.9 1331.9 1400.1 1566.3
#> 2016 1519.2 1155.8 1451.5 1451.0 1449.7 1596.1 1468.3 1293.9 1393.5 1497.4
#> 2017 1428.5 1092.4 1370.3 1522.6 1452.4 1557.2 1445.5 1303.1              
#>         Nov    Dec
#> 1982  522.9  889.3
#> 1983  568.7  963.7
#> 1984  623.2 1039.4
#> 1985  697.1 1187.5
#> 1986  700.9 1367.9
#> 1987  743.2 1460.1
#> 1988  870.3 1570.0
#> 1989  923.8 1569.4
#> 1990  931.5 1563.1
#> 1991  944.7 1600.3
#> 1992  968.6 1650.9
#> 1993  970.0 1710.5
#> 1994 1000.4 1817.5
#> 1995 1067.4 1857.2
#> 1996 1081.6 1837.6
#> 1997 1111.3 1879.1
#> 1998 1137.7 1975.7
#> 1999 1207.2 2069.6
#> 2000 1232.5 2177.5
#> 2001 1344.8 2269.5
#> 2002 1495.1 2338.6
#> 2003 1528.2 2424.2
#> 2004 1536.7 2500.8
#> 2005 1540.4 2536.0
#> 2006 1623.3 2611.1
#> 2007 1687.9 2747.0
#> 2008 1675.9 2886.1
#> 2009 1736.6 2795.1
#> 2010 1705.5 2752.4
#> 2011 1660.0 2730.5
#> 2012 1672.9 2753.3
#> 2013 1668.9 2725.5
#> 2014 1687.7 2756.9
#> 2015 1730.5 2913.6
#> 2016 1684.3 2850.4
#> 2017              
#> 
#> $model$variables
#> [1] "AO.220" "AO.219" "AO.277" "LS.400" "LS.280"
#> 
#> $model$X
#>        [,1] [,2] [,3] [,4] [,5]
#>   [1,]    0    0    0   -1   -1
#>   [2,]    0    0    0   -1   -1
#>   [3,]    0    0    0   -1   -1
#>   [4,]    0    0    0   -1   -1
#>   [5,]    0    0    0   -1   -1
#>   [6,]    0    0    0   -1   -1
#>   [7,]    0    0    0   -1   -1
#>   [8,]    0    0    0   -1   -1
#>   [9,]    0    0    0   -1   -1
#>  [10,]    0    0    0   -1   -1
#>  [11,]    0    0    0   -1   -1
#>  [12,]    0    0    0   -1   -1
#>  [13,]    0    0    0   -1   -1
#>  [14,]    0    0    0   -1   -1
#>  [15,]    0    0    0   -1   -1
#>  [16,]    0    0    0   -1   -1
#>  [17,]    0    0    0   -1   -1
#>  [18,]    0    0    0   -1   -1
#>  [19,]    0    0    0   -1   -1
#>  [20,]    0    0    0   -1   -1
#>  [21,]    0    0    0   -1   -1
#>  [22,]    0    0    0   -1   -1
#>  [23,]    0    0    0   -1   -1
#>  [24,]    0    0    0   -1   -1
#>  [25,]    0    0    0   -1   -1
#>  [26,]    0    0    0   -1   -1
#>  [27,]    0    0    0   -1   -1
#>  [28,]    0    0    0   -1   -1
#>  [29,]    0    0    0   -1   -1
#>  [30,]    0    0    0   -1   -1
#>  [31,]    0    0    0   -1   -1
#>  [32,]    0    0    0   -1   -1
#>  [33,]    0    0    0   -1   -1
#>  [34,]    0    0    0   -1   -1
#>  [35,]    0    0    0   -1   -1
#>  [36,]    0    0    0   -1   -1
#>  [37,]    0    0    0   -1   -1
#>  [38,]    0    0    0   -1   -1
#>  [39,]    0    0    0   -1   -1
#>  [40,]    0    0    0   -1   -1
#>  [41,]    0    0    0   -1   -1
#>  [42,]    0    0    0   -1   -1
#>  [43,]    0    0    0   -1   -1
#>  [44,]    0    0    0   -1   -1
#>  [45,]    0    0    0   -1   -1
#>  [46,]    0    0    0   -1   -1
#>  [47,]    0    0    0   -1   -1
#>  [48,]    0    0    0   -1   -1
#>  [49,]    0    0    0   -1   -1
#>  [50,]    0    0    0   -1   -1
#>  [51,]    0    0    0   -1   -1
#>  [52,]    0    0    0   -1   -1
#>  [53,]    0    0    0   -1   -1
#>  [54,]    0    0    0   -1   -1
#>  [55,]    0    0    0   -1   -1
#>  [56,]    0    0    0   -1   -1
#>  [57,]    0    0    0   -1   -1
#>  [58,]    0    0    0   -1   -1
#>  [59,]    0    0    0   -1   -1
#>  [60,]    0    0    0   -1   -1
#>  [61,]    0    0    0   -1   -1
#>  [62,]    0    0    0   -1   -1
#>  [63,]    0    0    0   -1   -1
#>  [64,]    0    0    0   -1   -1
#>  [65,]    0    0    0   -1   -1
#>  [66,]    0    0    0   -1   -1
#>  [67,]    0    0    0   -1   -1
#>  [68,]    0    0    0   -1   -1
#>  [69,]    0    0    0   -1   -1
#>  [70,]    0    0    0   -1   -1
#>  [71,]    0    0    0   -1   -1
#>  [72,]    0    0    0   -1   -1
#>  [73,]    0    0    0   -1   -1
#>  [74,]    0    0    0   -1   -1
#>  [75,]    0    0    0   -1   -1
#>  [76,]    0    0    0   -1   -1
#>  [77,]    0    0    0   -1   -1
#>  [78,]    0    0    0   -1   -1
#>  [79,]    0    0    0   -1   -1
#>  [80,]    0    0    0   -1   -1
#>  [81,]    0    0    0   -1   -1
#>  [82,]    0    0    0   -1   -1
#>  [83,]    0    0    0   -1   -1
#>  [84,]    0    0    0   -1   -1
#>  [85,]    0    0    0   -1   -1
#>  [86,]    0    0    0   -1   -1
#>  [87,]    0    0    0   -1   -1
#>  [88,]    0    0    0   -1   -1
#>  [89,]    0    0    0   -1   -1
#>  [90,]    0    0    0   -1   -1
#>  [91,]    0    0    0   -1   -1
#>  [92,]    0    0    0   -1   -1
#>  [93,]    0    0    0   -1   -1
#>  [94,]    0    0    0   -1   -1
#>  [95,]    0    0    0   -1   -1
#>  [96,]    0    0    0   -1   -1
#>  [97,]    0    0    0   -1   -1
#>  [98,]    0    0    0   -1   -1
#>  [99,]    0    0    0   -1   -1
#> [100,]    0    0    0   -1   -1
#> [101,]    0    0    0   -1   -1
#> [102,]    0    0    0   -1   -1
#> [103,]    0    0    0   -1   -1
#> [104,]    0    0    0   -1   -1
#> [105,]    0    0    0   -1   -1
#> [106,]    0    0    0   -1   -1
#> [107,]    0    0    0   -1   -1
#> [108,]    0    0    0   -1   -1
#> [109,]    0    0    0   -1   -1
#> [110,]    0    0    0   -1   -1
#> [111,]    0    0    0   -1   -1
#> [112,]    0    0    0   -1   -1
#> [113,]    0    0    0   -1   -1
#> [114,]    0    0    0   -1   -1
#> [115,]    0    0    0   -1   -1
#> [116,]    0    0    0   -1   -1
#> [117,]    0    0    0   -1   -1
#> [118,]    0    0    0   -1   -1
#> [119,]    0    0    0   -1   -1
#> [120,]    0    0    0   -1   -1
#> [121,]    0    0    0   -1   -1
#> [122,]    0    0    0   -1   -1
#> [123,]    0    0    0   -1   -1
#> [124,]    0    0    0   -1   -1
#> [125,]    0    0    0   -1   -1
#> [126,]    0    0    0   -1   -1
#> [127,]    0    0    0   -1   -1
#> [128,]    0    0    0   -1   -1
#> [129,]    0    0    0   -1   -1
#> [130,]    0    0    0   -1   -1
#> [131,]    0    0    0   -1   -1
#> [132,]    0    0    0   -1   -1
#> [133,]    0    0    0   -1   -1
#> [134,]    0    0    0   -1   -1
#> [135,]    0    0    0   -1   -1
#> [136,]    0    0    0   -1   -1
#> [137,]    0    0    0   -1   -1
#> [138,]    0    0    0   -1   -1
#> [139,]    0    0    0   -1   -1
#> [140,]    0    0    0   -1   -1
#> [141,]    0    0    0   -1   -1
#> [142,]    0    0    0   -1   -1
#> [143,]    0    0    0   -1   -1
#> [144,]    0    0    0   -1   -1
#> [145,]    0    0    0   -1   -1
#> [146,]    0    0    0   -1   -1
#> [147,]    0    0    0   -1   -1
#> [148,]    0    0    0   -1   -1
#> [149,]    0    0    0   -1   -1
#> [150,]    0    0    0   -1   -1
#> [151,]    0    0    0   -1   -1
#> [152,]    0    0    0   -1   -1
#> [153,]    0    0    0   -1   -1
#> [154,]    0    0    0   -1   -1
#> [155,]    0    0    0   -1   -1
#> [156,]    0    0    0   -1   -1
#> [157,]    0    0    0   -1   -1
#> [158,]    0    0    0   -1   -1
#> [159,]    0    0    0   -1   -1
#> [160,]    0    0    0   -1   -1
#> [161,]    0    0    0   -1   -1
#> [162,]    0    0    0   -1   -1
#> [163,]    0    0    0   -1   -1
#> [164,]    0    0    0   -1   -1
#> [165,]    0    0    0   -1   -1
#> [166,]    0    0    0   -1   -1
#> [167,]    0    0    0   -1   -1
#> [168,]    0    0    0   -1   -1
#> [169,]    0    0    0   -1   -1
#> [170,]    0    0    0   -1   -1
#> [171,]    0    0    0   -1   -1
#> [172,]    0    0    0   -1   -1
#> [173,]    0    0    0   -1   -1
#> [174,]    0    0    0   -1   -1
#> [175,]    0    0    0   -1   -1
#> [176,]    0    0    0   -1   -1
#> [177,]    0    0    0   -1   -1
#> [178,]    0    0    0   -1   -1
#> [179,]    0    0    0   -1   -1
#> [180,]    0    0    0   -1   -1
#> [181,]    0    0    0   -1   -1
#> [182,]    0    0    0   -1   -1
#> [183,]    0    0    0   -1   -1
#> [184,]    0    0    0   -1   -1
#> [185,]    0    0    0   -1   -1
#> [186,]    0    0    0   -1   -1
#> [187,]    0    0    0   -1   -1
#> [188,]    0    0    0   -1   -1
#> [189,]    0    0    0   -1   -1
#> [190,]    0    0    0   -1   -1
#> [191,]    0    0    0   -1   -1
#> [192,]    0    0    0   -1   -1
#> [193,]    0    0    0   -1   -1
#> [194,]    0    0    0   -1   -1
#> [195,]    0    0    0   -1   -1
#> [196,]    0    0    0   -1   -1
#> [197,]    0    0    0   -1   -1
#> [198,]    0    0    0   -1   -1
#> [199,]    0    0    0   -1   -1
#> [200,]    0    0    0   -1   -1
#> [201,]    0    0    0   -1   -1
#> [202,]    0    0    0   -1   -1
#> [203,]    0    0    0   -1   -1
#> [204,]    0    0    0   -1   -1
#> [205,]    0    0    0   -1   -1
#> [206,]    0    0    0   -1   -1
#> [207,]    0    0    0   -1   -1
#> [208,]    0    0    0   -1   -1
#> [209,]    0    0    0   -1   -1
#> [210,]    0    0    0   -1   -1
#> [211,]    0    0    0   -1   -1
#> [212,]    0    0    0   -1   -1
#> [213,]    0    0    0   -1   -1
#> [214,]    0    0    0   -1   -1
#> [215,]    0    0    0   -1   -1
#> [216,]    0    0    0   -1   -1
#> [217,]    0    0    0   -1   -1
#> [218,]    0    0    0   -1   -1
#> [219,]    0    1    0   -1   -1
#> [220,]    1    0    0   -1   -1
#> [221,]    0    0    0   -1   -1
#> [222,]    0    0    0   -1   -1
#> [223,]    0    0    0   -1   -1
#> [224,]    0    0    0   -1   -1
#> [225,]    0    0    0   -1   -1
#> [226,]    0    0    0   -1   -1
#> [227,]    0    0    0   -1   -1
#> [228,]    0    0    0   -1   -1
#> [229,]    0    0    0   -1   -1
#> [230,]    0    0    0   -1   -1
#> [231,]    0    0    0   -1   -1
#> [232,]    0    0    0   -1   -1
#> [233,]    0    0    0   -1   -1
#> [234,]    0    0    0   -1   -1
#> [235,]    0    0    0   -1   -1
#> [236,]    0    0    0   -1   -1
#> [237,]    0    0    0   -1   -1
#> [238,]    0    0    0   -1   -1
#> [239,]    0    0    0   -1   -1
#> [240,]    0    0    0   -1   -1
#> [241,]    0    0    0   -1   -1
#> [242,]    0    0    0   -1   -1
#> [243,]    0    0    0   -1   -1
#> [244,]    0    0    0   -1   -1
#> [245,]    0    0    0   -1   -1
#> [246,]    0    0    0   -1   -1
#> [247,]    0    0    0   -1   -1
#> [248,]    0    0    0   -1   -1
#> [249,]    0    0    0   -1   -1
#> [250,]    0    0    0   -1   -1
#> [251,]    0    0    0   -1   -1
#> [252,]    0    0    0   -1   -1
#> [253,]    0    0    0   -1   -1
#> [254,]    0    0    0   -1   -1
#> [255,]    0    0    0   -1   -1
#> [256,]    0    0    0   -1   -1
#> [257,]    0    0    0   -1   -1
#> [258,]    0    0    0   -1   -1
#> [259,]    0    0    0   -1   -1
#> [260,]    0    0    0   -1   -1
#> [261,]    0    0    0   -1   -1
#> [262,]    0    0    0   -1   -1
#> [263,]    0    0    0   -1   -1
#> [264,]    0    0    0   -1   -1
#> [265,]    0    0    0   -1   -1
#> [266,]    0    0    0   -1   -1
#> [267,]    0    0    0   -1   -1
#> [268,]    0    0    0   -1   -1
#> [269,]    0    0    0   -1   -1
#> [270,]    0    0    0   -1   -1
#> [271,]    0    0    0   -1   -1
#> [272,]    0    0    0   -1   -1
#> [273,]    0    0    0   -1   -1
#> [274,]    0    0    0   -1   -1
#> [275,]    0    0    0   -1   -1
#> [276,]    0    0    0   -1   -1
#> [277,]    0    0    1   -1   -1
#> [278,]    0    0    0   -1   -1
#> [279,]    0    0    0   -1   -1
#> [280,]    0    0    0   -1    0
#> [281,]    0    0    0   -1    0
#> [282,]    0    0    0   -1    0
#> [283,]    0    0    0   -1    0
#> [284,]    0    0    0   -1    0
#> [285,]    0    0    0   -1    0
#> [286,]    0    0    0   -1    0
#> [287,]    0    0    0   -1    0
#> [288,]    0    0    0   -1    0
#> [289,]    0    0    0   -1    0
#> [290,]    0    0    0   -1    0
#> [291,]    0    0    0   -1    0
#> [292,]    0    0    0   -1    0
#> [293,]    0    0    0   -1    0
#> [294,]    0    0    0   -1    0
#> [295,]    0    0    0   -1    0
#> [296,]    0    0    0   -1    0
#> [297,]    0    0    0   -1    0
#> [298,]    0    0    0   -1    0
#> [299,]    0    0    0   -1    0
#> [300,]    0    0    0   -1    0
#> [301,]    0    0    0   -1    0
#> [302,]    0    0    0   -1    0
#> [303,]    0    0    0   -1    0
#> [304,]    0    0    0   -1    0
#> [305,]    0    0    0   -1    0
#> [306,]    0    0    0   -1    0
#> [307,]    0    0    0   -1    0
#> [308,]    0    0    0   -1    0
#> [309,]    0    0    0   -1    0
#> [310,]    0    0    0   -1    0
#> [311,]    0    0    0   -1    0
#> [312,]    0    0    0   -1    0
#> [313,]    0    0    0   -1    0
#> [314,]    0    0    0   -1    0
#> [315,]    0    0    0   -1    0
#> [316,]    0    0    0   -1    0
#> [317,]    0    0    0   -1    0
#> [318,]    0    0    0   -1    0
#> [319,]    0    0    0   -1    0
#> [320,]    0    0    0   -1    0
#> [321,]    0    0    0   -1    0
#> [322,]    0    0    0   -1    0
#> [323,]    0    0    0   -1    0
#> [324,]    0    0    0   -1    0
#> [325,]    0    0    0   -1    0
#> [326,]    0    0    0   -1    0
#> [327,]    0    0    0   -1    0
#> [328,]    0    0    0   -1    0
#> [329,]    0    0    0   -1    0
#> [330,]    0    0    0   -1    0
#> [331,]    0    0    0   -1    0
#> [332,]    0    0    0   -1    0
#> [333,]    0    0    0   -1    0
#> [334,]    0    0    0   -1    0
#> [335,]    0    0    0   -1    0
#> [336,]    0    0    0   -1    0
#> [337,]    0    0    0   -1    0
#> [338,]    0    0    0   -1    0
#> [339,]    0    0    0   -1    0
#> [340,]    0    0    0   -1    0
#> [341,]    0    0    0   -1    0
#> [342,]    0    0    0   -1    0
#> [343,]    0    0    0   -1    0
#> [344,]    0    0    0   -1    0
#> [345,]    0    0    0   -1    0
#> [346,]    0    0    0   -1    0
#> [347,]    0    0    0   -1    0
#> [348,]    0    0    0   -1    0
#> [349,]    0    0    0   -1    0
#> [350,]    0    0    0   -1    0
#> [351,]    0    0    0   -1    0
#> [352,]    0    0    0   -1    0
#> [353,]    0    0    0   -1    0
#> [354,]    0    0    0   -1    0
#> [355,]    0    0    0   -1    0
#> [356,]    0    0    0   -1    0
#> [357,]    0    0    0   -1    0
#> [358,]    0    0    0   -1    0
#> [359,]    0    0    0   -1    0
#> [360,]    0    0    0   -1    0
#> [361,]    0    0    0   -1    0
#> [362,]    0    0    0   -1    0
#> [363,]    0    0    0   -1    0
#> [364,]    0    0    0   -1    0
#> [365,]    0    0    0   -1    0
#> [366,]    0    0    0   -1    0
#> [367,]    0    0    0   -1    0
#> [368,]    0    0    0   -1    0
#> [369,]    0    0    0   -1    0
#> [370,]    0    0    0   -1    0
#> [371,]    0    0    0   -1    0
#> [372,]    0    0    0   -1    0
#> [373,]    0    0    0   -1    0
#> [374,]    0    0    0   -1    0
#> [375,]    0    0    0   -1    0
#> [376,]    0    0    0   -1    0
#> [377,]    0    0    0   -1    0
#> [378,]    0    0    0   -1    0
#> [379,]    0    0    0   -1    0
#> [380,]    0    0    0   -1    0
#> [381,]    0    0    0   -1    0
#> [382,]    0    0    0   -1    0
#> [383,]    0    0    0   -1    0
#> [384,]    0    0    0   -1    0
#> [385,]    0    0    0   -1    0
#> [386,]    0    0    0   -1    0
#> [387,]    0    0    0   -1    0
#> [388,]    0    0    0   -1    0
#> [389,]    0    0    0   -1    0
#> [390,]    0    0    0   -1    0
#> [391,]    0    0    0   -1    0
#> [392,]    0    0    0   -1    0
#> [393,]    0    0    0   -1    0
#> [394,]    0    0    0   -1    0
#> [395,]    0    0    0   -1    0
#> [396,]    0    0    0   -1    0
#> [397,]    0    0    0   -1    0
#> [398,]    0    0    0   -1    0
#> [399,]    0    0    0   -1    0
#> [400,]    0    0    0    0    0
#> [401,]    0    0    0    0    0
#> [402,]    0    0    0    0    0
#> [403,]    0    0    0    0    0
#> [404,]    0    0    0    0    0
#> [405,]    0    0    0    0    0
#> [406,]    0    0    0    0    0
#> [407,]    0    0    0    0    0
#> [408,]    0    0    0    0    0
#> [409,]    0    0    0    0    0
#> [410,]    0    0    0    0    0
#> [411,]    0    0    0    0    0
#> [412,]    0    0    0    0    0
#> [413,]    0    0    0    0    0
#> [414,]    0    0    0    0    0
#> [415,]    0    0    0    0    0
#> [416,]    0    0    0    0    0
#> [417,]    0    0    0    0    0
#> [418,]    0    0    0    0    0
#> [419,]    0    0    0    0    0
#> [420,]    0    0    0    0    0
#> [421,]    0    0    0    0    0
#> [422,]    0    0    0    0    0
#> [423,]    0    0    0    0    0
#> [424,]    0    0    0    0    0
#> [425,]    0    0    0    0    0
#> 
#> $model$b
#> [1] -210.91745  199.66569 -192.39785   84.32270  -77.78577
#> 
#> $model$bcov
#>               [,1]          [,2]          [,3]          [,4]          [,5]
#> [1,]  1.258544e+03  5.076022e+01 -8.906666e-01 -1.197400e-05  -0.827996523
#> [2,]  5.076022e+01  1.258545e+03 -9.422003e-01 -1.454505e-05  -1.048251277
#> [3,] -8.906666e-01 -9.422003e-01  1.322416e+03  5.431398e-04 155.010519468
#> [4,] -1.197400e-05 -1.454505e-05  5.431398e-04  3.654102e+02   0.004446791
#> [5,] -8.279965e-01 -1.048251e+00  1.550105e+02  4.446791e-03 364.501084888
#> 
#> $model$linearized
#>   [1]  466.6369  509.1369  450.3369  465.6369  444.9369  471.6369  459.2369
#>   [8]  529.4369  895.8369  385.7369  384.5369  478.6369  509.9369  517.1369
#>  [15]  468.9369  474.8369  464.7369  489.2369  491.8369  575.2369  970.2369
#>  [22]  421.2369  421.0369  491.2369  493.8369  604.4369  506.9369  549.9369
#>  [29]  509.9369  529.3369  563.1369  629.7369 1045.9369  522.8369  459.0369
#>  [36]  532.3369  594.2369  706.8369  568.3369  609.3369  589.0369  569.6369
#>  [43]  643.6369  703.6369 1194.0369  577.0369  484.7369  553.9369  600.8369
#>  [50]  758.1369  559.9369  669.7369  587.6369  668.4369  672.1369  707.4369
#>  [57] 1374.4369  620.4369  519.7369  606.4369  680.6369  720.5369  677.0369
#>  [64]  727.4369  608.1369  678.8369  715.6369  749.7369 1466.6369  637.5369
#>  [71]  557.6369  684.6369  722.2369  747.3369  728.5369  690.0369  657.4369
#>  [78]  729.8369  736.1369  876.8369 1576.5369  638.0369  558.5369  725.5369
#>  [85]  704.1369  771.3369  792.8369  721.6369  730.3369  764.4369  758.2369
#>  [92]  930.3369 1575.9369  684.7369  592.7369  733.3369  750.6369  822.0369
#>  [99]  838.9369  716.8369  765.9369  747.6369  793.1369  938.0369 1569.6369
#> [106]  700.5369  611.2369  725.7369  754.7369  834.7369  753.4369  801.0369
#> [113]  776.9369  748.0369  865.1369  951.2369 1606.8369  746.5369  672.4369
#> [120]  708.0369  837.9369  885.1369  832.5369  794.7369  730.1369  826.3369
#> [127]  909.0369  975.1369 1657.4369  768.6369  649.5369  760.6369  847.2369
#> [134]  913.1369  893.6369  778.0369  735.2369  851.2369  893.4369  976.5369
#> [141] 1717.0369  752.2369  670.9369  828.0369  838.2369  914.5369  919.1369
#> [148]  789.4369  805.3369  893.5369  941.1369 1006.9369 1824.0369  758.9369
#> [155]  689.0369  817.7369  912.5369  933.7369  913.3369  887.1369  880.4369
#> [162]  863.3369  927.1369 1073.9369 1863.7369  839.6369  743.6369  818.5369
#> [169]  901.7369  969.3369  915.1369  914.5369  895.4369  840.2369  940.2369
#> [176] 1088.1369 1844.1369  847.4369  733.9369  864.4369  855.5369 1001.3369
#> [183]  836.7369  977.6369  842.5369  945.6369  983.4369 1117.8369 1885.6369
#> [190]  923.8369  722.7369  829.4369  976.6369  976.7369  855.9369 1048.8369
#> [197]  876.4369  945.9369 1027.8369 1144.2369 1982.2369  948.5369  744.9369
#> [204]  909.7369  959.7369 1017.7369  900.9369 1061.0369  906.0369 1008.8369
#> [211] 1050.2369 1213.7369 2076.1369  931.4369  804.7369  908.4369 1031.2369
#> [218] 1058.8369  972.3712 1076.7544 1015.7369 1061.1369 1076.9369 1239.0369
#> [225] 2184.0369  978.4369  821.1369 1024.0369 1045.7369 1130.0369 1031.4369
#> [232] 1107.3369  969.5369 1019.4369 1138.5369 1351.3369 2276.0369 1034.4369
#> [239]  847.9369 1050.4369 1081.8369 1197.4369 1149.5369 1082.2369 1072.4369
#> [246] 1066.6369 1217.9369 1501.6369 2345.1369 1105.8369  907.0369 1099.2369
#> [253] 1228.9369 1243.6369 1244.4369 1188.5369 1107.7369 1204.7369 1322.6369
#> [260] 1534.7369 2430.7369 1189.4369  996.3369 1137.9369 1283.6369 1286.8369
#> [267] 1390.6369 1312.4369 1173.3369 1324.4369 1364.8369 1543.2369 2507.3369
#> [274] 1252.8369 1043.8369 1307.3369 1352.6348 1270.7369 1460.7369 1374.4227
#> [281] 1295.0227 1362.1227 1398.7227 1624.7227 2620.3227 1278.0227 1122.0227
#> [288] 1288.8227 1432.9227 1351.9227 1513.3227 1496.3227 1323.5227 1303.4227
#> [295] 1428.9227 1707.6227 2695.4227 1351.6227 1131.3227 1415.9227 1386.9227
#> [302] 1449.4227 1575.8227 1546.6227 1399.8227 1437.6227 1524.9227 1772.2227
#> [309] 2831.3227 1482.1227 1224.8227 1436.0227 1480.9227 1505.4227 1485.9227
#> [316] 1666.6227 1352.7227 1467.6227 1536.7227 1760.2227 2970.4227 1535.3227
#> [323] 1140.9227 1471.2227 1593.4227 1603.7227 1584.8227 1655.0227 1425.8227
#> [330] 1484.2227 1618.6227 1820.9227 2879.4227 1553.4227 1196.2227 1464.2227
#> [337] 1474.0227 1511.5227 1635.7227 1665.3227 1408.3227 1506.3227 1549.2227
#> [344] 1789.8227 2836.7227 1496.9227 1201.8227 1405.9227 1556.9227 1493.2227
#> [351] 1556.2227 1616.8227 1377.8227 1430.0227 1489.0227 1744.3227 2814.8227
#> [358] 1446.7227 1216.0227 1433.5227 1475.5227 1541.2227 1700.7227 1507.7227
#> [365] 1443.3227 1452.1227 1526.9227 1757.2227 2837.6227 1481.7227 1197.9227
#> [372] 1481.6227 1423.4227 1526.2227 1621.7227 1474.9227 1421.5227 1443.7227
#> [379] 1547.6227 1753.2227 2809.8227 1535.3227 1149.2227 1377.5227 1527.2227
#> [386] 1496.1227 1545.9227 1585.9227 1338.5227 1440.7227 1563.0227 1772.0227
#> [393] 2841.2227 1555.5227 1138.1227 1451.5227 1526.5227 1513.0227 1565.2227
#> [400] 1540.9000 1331.9000 1400.1000 1566.3000 1730.5000 2913.6000 1519.2000
#> [407] 1155.8000 1451.5000 1451.0000 1449.7000 1596.1000 1468.3000 1293.9000
#> [414] 1393.5000 1497.4000 1684.3000 2850.4000 1428.5000 1092.4000 1370.3000
#> [421] 1522.6000 1452.4000 1557.2000 1445.5000 1303.1000
#> 
#> 
#> $likelihood
#> $likelihood$initial
#> $likelihood$initial$ll
#> [1] -2218.964
#> 
#> $likelihood$initial$ssq
#> [1] 1139531
#> 
#> $likelihood$initial$nobs
#> [1] 425
#> 
#> $likelihood$initial$neffective
#> [1] -1
#> 
#> $likelihood$initial$nparams
#> [1] 3
#> 
#> $likelihood$initial$df
#> [1] 409
#> 
#> $likelihood$initial$aic
#> [1] 4443.928
#> 
#> $likelihood$initial$aicc
#> [1] 4443.987
#> 
#> $likelihood$initial$bic
#> [1] 4455.991
#> 
#> $likelihood$initial$bic2
#> [1] 10.81551
#> 
#> $likelihood$initial$bicc
#> [1] 7.954332
#> 
#> $likelihood$initial$hannanquinn
#> [1] 4448.7
#> 
#> 
#> $likelihood$final
#> $likelihood$final$ll
#> [1] -2167.948
#> 
#> $likelihood$final$ssq
#> [1] 892303.8
#> 
#> $likelihood$final$nobs
#> [1] 425
#> 
#> $likelihood$final$neffective
#> [1] -1
#> 
#> $likelihood$final$nparams
#> [1] 6
#> 
#> $likelihood$final$df
#> [1] 406
#> 
#> $likelihood$final$aic
#> [1] 4347.896
#> 
#> $likelihood$final$aicc
#> [1] 4348.104
#> 
#> $likelihood$final$bic
#> [1] 4372.023
#> 
#> $likelihood$final$bic2
#> [1] 10.61171
#> 
#> $likelihood$final$bicc
#> [1] 7.753609
#> 
#> $likelihood$final$hannanquinn
#> [1] 4357.439
#> 
#> 
#> 
#> attr(,"class")
#> [1] "JD3_REGARIMA_OUTLIERS"