Seasonal Adjustment with X13-ARIMA

Seasonal Adjustment with X13-ARIMA

Usage

x13(
  ts,
  spec = c("rsa4", "rsa0", "rsa1", "rsa2c", "rsa3", "rsa5c"),
  context = NULL,
  userdefined = NULL
)

x13_fast(
  ts,
  spec = c("rsa4", "rsa0", "rsa1", "rsa2c", "rsa3", "rsa5c"),
  context = NULL,
  userdefined = NULL
)

jx13(
  ts,
  spec = c("rsa4", "rsa0", "rsa1", "rsa2c", "rsa3", "rsa5c"),
  context = NULL,
  userdefined = NULL
)

Arguments

ts

an univariate time series.

spec

the model specification. Can be either the name of a predefined specification or a user-defined specification.

context

list of external regressors (calendar or other) to be used for estimation

userdefined

a vector containing additional output variables (see x13_dictionary()).

Value

the x13() function returns a list with the results, the estimation specification and the result specification, while x13_fast() is a faster function that only returns the results. The jx13() functions only returns results in a java object which will allow to customize outputs in other packages (use rjd3toolkit::dictionary() to get the list of variables and rjd3toolkit::result() to get a specific variable). In the estimation functions x13() and x13_fast() you can directly use a specification name (string). If you want to customize a specification you have to create a specification object first.

Examples

y = rjd3toolkit::ABS$X0.2.09.10.M
x13_fast(y,"rsa3")
#> RegARIMA
#> Log-transformation: yes 
#> SARIMA model:  (2,1,1) (0,1,1)
#> 
#> Coefficients
#>           Estimate Std. Error  T-stat
#> phi(1)     0.38042    0.06517   5.838
#> phi(2)     0.21398    0.06040   3.543
#> theta(1)  -0.70246    0.05146 -13.650
#> btheta(1) -0.56259    0.06648  -8.462
#> 
#> Regression model:
#>                 Estimate Std. Error T-stat
#> TC (2000-06-01)  0.15476    0.03158  4.900
#> AO (2000-07-01) -0.29505    0.04425 -6.668
#> Number of observations:  425 
#> Number of effective observations:  412 
#> Number of parameters:  7 
#> 
#> Loglikelihood:  720.2196 
#> Adjusted loglikelihood:  -2147.407 
#> 
#> Standard error of the regression (ML estimate):  0.04181753 
#> AIC:  4308.814 
#> AICC:  4309.092 
#> BIC:  4336.962 
#> 
#> 
#> Decomposition
#> Monitoring and Quality Assessment Statistics: 
#>     M stats
#> m1    0.095
#> m2    0.097
#> m3    2.384
#> m4    0.403
#> m5    1.940
#> m6    0.179
#> m7    0.053
#> m8    0.176
#> m9    0.047
#> m10   0.157
#> m11   0.154
#> q     0.560
#> qm2   0.617
#> 
#> Final filters: 
#> Seasonal filter:  
#> Trend filter: 23 terms Henderson moving average
#> 
#> Diagnostics
#> Relative contribution of the components to the stationary
#> portion of the variance in the original series,
#> after the removal of the long term trend (in %)
#> 
#>            Component
#>  cycle        13.259
#>  seasonal     86.854
#>  irregular     0.970
#>  calendar      0.000
#>  others        0.299
#>  total       101.383
#> 
#> Residual seasonality tests
#>                 P.value
#>  seas.ftest.i     0.994
#>  seas.ftest.sa    0.999
#>  seas.qstest.i    0.045
#>  seas.qstest.sa   0.035
#>  td.ftest.i       0.000
#>  td.ftest.sa      0.000
#> 
#> 
#> Final
#> Last values
#>          series       sa    trend      seas       irr
#> Sep 2016 1393.5 1558.700 1547.466 0.8940140 1.0072601
#> Oct 2016 1497.4 1539.619 1543.359 0.9725785 0.9975766
#> Nov 2016 1684.3 1540.423 1541.287 1.0934011 0.9994390
#> Dec 2016 2850.4 1561.106 1541.060 1.8258848 1.0130078
#> Jan 2017 1428.5 1513.071 1542.061 0.9441065 0.9812005
#> Feb 2017 1092.4 1538.709 1543.697 0.7099456 0.9967689
#> Mar 2017 1370.3 1532.681 1545.484 0.8940545 0.9917157
#> Apr 2017 1522.6 1609.073 1547.232 0.9462594 1.0399684
#> May 2017 1452.4 1563.830 1548.782 0.9287456 1.0097158
#> Jun 2017 1557.2 1561.361 1549.791 0.9973351 1.0074658
#> Jul 2017 1445.5 1519.197 1550.206 0.9514898 0.9799965
#> Aug 2017 1303.1 1554.064 1550.306 0.8385112 1.0024242
x13(y,"rsa5c")
#> RegARIMA
#> Log-transformation: yes 
#> SARIMA model:  (0,1,1) (1,1,1)
#> 
#> Coefficients
#>           Estimate Std. Error T-stat
#> theta(1)  -0.81551    0.06932 -11.76
#> bphi(1)   -0.43410    0.02650 -16.38
#> btheta(1) -0.82463    0.04458 -18.50
#> 
#> Regression model:
#>                  Estimate Std. Error T-stat
#> monday          -0.009156   0.003278 -2.793
#> tuesday          0.004523   0.003316  1.364
#> wednesday       -0.001181   0.003285 -0.360
#> thursday         0.013349   0.003304  4.040
#> friday          -0.001501   0.003313 -0.453
#> saturday         0.014993   0.003293  4.553
#> easter           0.051061   0.006629  7.703
#> TC (2000-06-01)  0.162337   0.026546  6.115
#> AO (2000-07-01) -0.306371   0.032121 -9.538
#> Number of observations:  425 
#> Number of effective observations:  412 
#> Number of parameters:  13 
#> 
#> Loglikelihood:  794.8103 
#> Adjusted loglikelihood:  -2072.816 
#> 
#> Standard error of the regression (ML estimate):  0.03486636 
#> AIC:  4171.633 
#> AICC:  4172.548 
#> BIC:  4223.906 
#> 
#> 
#> Decomposition
#> Monitoring and Quality Assessment Statistics: 
#>     M stats
#> m1    0.054
#> m2    0.049
#> m3    1.765
#> m4    0.089
#> m5    1.416
#> m6    0.012
#> m7    0.052
#> m8    0.160
#> m9    0.050
#> m10   0.124
#> m11   0.122
#> q     0.385
#> qm2   0.427
#> 
#> Final filters: 
#> Seasonal filter:  
#> Trend filter: 23 terms Henderson moving average
#> 
#> Diagnostics
#> Relative contribution of the components to the stationary
#> portion of the variance in the original series,
#> after the removal of the long term trend (in %)
#> 
#>            Component
#>  cycle        13.437
#>  seasonal     86.536
#>  irregular     0.489
#>  calendar      0.599
#>  others        0.323
#>  total       101.383
#> 
#> Residual seasonality tests
#>                 P.value
#>  seas.ftest.i     0.977
#>  seas.ftest.sa    0.986
#>  seas.qstest.i    0.010
#>  seas.qstest.sa   0.004
#>  td.ftest.i       0.993
#>  td.ftest.sa      0.996
#> 
#> 
#> Final
#> Last values
#>          series       sa    trend      seas       irr
#> Sep 2016 1393.5 1544.956 1548.898 0.9019675 0.9974547
#> Oct 2016 1497.4 1569.404 1547.139 0.9541202 1.0143910
#> Nov 2016 1684.3 1523.294 1546.637 1.1056957 0.9849077
#> Dec 2016 2850.4 1536.881 1546.937 1.8546651 0.9934994
#> Jan 2017 1428.5 1547.497 1547.656 0.9231038 0.9998970
#> Feb 2017 1092.4 1545.588 1548.350 0.7067860 0.9982164
#> Mar 2017 1370.3 1556.927 1548.747 0.8801312 1.0052818
#> Apr 2017 1522.6 1587.924 1548.518 0.9588622 1.0254473
#> May 2017 1452.4 1555.232 1547.335 0.9338798 1.0051038
#> Jun 2017 1557.2 1539.880 1545.271 1.0112477 0.9965115
#> Jul 2017 1445.5 1532.110 1542.807 0.9434701 0.9930664
#> Aug 2017 1303.1 1524.546 1540.399 0.8547462 0.9897086
regarima_fast(y,"rg0")
#> Log-transformation: no 
#> SARIMA model:  (0,1,1) (0,1,1)
#> 
#> Coefficients
#>           Estimate Std. Error  T-stat
#> theta(1)  -0.87640    0.02290 -38.275
#> btheta(1) -0.38755    0.05071  -7.642
#> 
#> No regression variables
#> Number of observations:  425 
#> Number of effective observations:  412 
#> Number of parameters:  3 
#> 
#> Loglikelihood:  -2218.964 
#> Standard error of the regression (ML estimate):  52.59136 
#> AIC:  4443.928 
#> AICC:  4443.987 
#> BIC:  4455.991 
#> 
regarima(y,"rg3")
#> Log-transformation: yes 
#> SARIMA model:  (2,1,1) (0,1,1)
#> 
#> Coefficients
#>           Estimate Std. Error  T-stat
#> phi(1)     0.38042    0.06517   5.838
#> phi(2)     0.21398    0.06040   3.543
#> theta(1)  -0.70246    0.05146 -13.650
#> btheta(1) -0.56259    0.06648  -8.462
#> 
#> Regression model:
#>                 Estimate Std. Error T-stat
#> TC (2000-06-01)  0.15476    0.03158  4.900
#> AO (2000-07-01) -0.29505    0.04425 -6.668
#> Number of observations:  425 
#> Number of effective observations:  412 
#> Number of parameters:  7 
#> 
#> Loglikelihood:  720.2196 
#> Adjusted loglikelihood:  -2147.407 
#> 
#> Standard error of the regression (ML estimate):  0.04181753 
#> AIC:  4308.814 
#> AICC:  4309.092 
#> BIC:  4336.962 
#> 

sp = x13_spec("rsa5c")
sp = rjd3toolkit::add_outlier(sp,
                 type = c("AO"), c("2015-01-01", "2010-01-01"))
sp =  rjd3toolkit::set_transform(
   rjd3toolkit::set_tradingdays(
     rjd3toolkit::set_easter(sp, enabled = FALSE),
    option = "workingdays"
  ),
  fun = "None"
)
x13(y,spec=sp)
#> RegARIMA
#> Log-transformation: no 
#> SARIMA model:  (3,1,1) (0,1,1)
#> 
#> Coefficients
#>           Estimate Std. Error  T-stat
#> phi(1)     0.16557    0.06461   2.563
#> phi(2)     0.08780    0.06435   1.364
#> phi(3)    -0.11287    0.06047  -1.867
#> theta(1)  -0.86076    0.03927 -21.918
#> btheta(1) -0.22918    0.05193  -4.413
#> 
#> Regression model:
#>                  Estimate Std. Error T-stat
#> td                 1.0015     0.8098  1.237
#> lp                29.6392    11.5469  2.567
#> AO (2010-01-01)   37.0476    34.4300  1.076
#> AO (2015-01-01)   27.9946    35.0234  0.799
#> AO (2000-06-01)  199.7536    34.5756  5.777
#> AO (2000-07-01) -194.6887    34.6240 -5.623
#> LS (2005-04-01)  -82.3062    17.4210 -4.725
#> LS (2015-07-01)   81.3334    18.0924  4.495
#> Number of observations:  425 
#> Number of effective observations:  412 
#> Number of parameters:  14 
#> 
#> Loglikelihood:  -2159.902 
#> Standard error of the regression (ML estimate):  45.632 
#> AIC:  4347.804 
#> AICC:  4348.862 
#> BIC:  4404.099 
#> 
#> 
#> Decomposition
#> Monitoring and Quality Assessment Statistics: 
#>     M stats
#> m1    0.043
#> m2    0.033
#> m3    1.909
#> m4    0.447
#> m5    1.607
#> m6    0.399
#> m7    0.167
#> m8    0.298
#> m9    0.145
#> m10   0.308
#> m11   0.293
#> q     0.552
#> qm2   0.624
#> 
#> Final filters: 
#> Seasonal filter:  
#> Trend filter: 23 terms Henderson moving average
#> 
#> Diagnostics
#> Relative contribution of the components to the stationary
#> portion of the variance in the original series,
#> after the removal of the long term trend (in %)
#> 
#>            Component
#>  cycle         6.178
#>  seasonal     94.402
#>  irregular     0.325
#>  calendar      0.020
#>  others        1.504
#>  total       102.429
#> 
#> Residual seasonality tests
#>                 P.value
#>  seas.ftest.i     0.902
#>  seas.ftest.sa    0.996
#>  seas.qstest.i    0.038
#>  seas.qstest.sa   0.138
#>  td.ftest.i       0.000
#>  td.ftest.sa      0.000
#> 
#> 
#> Final
#> Last values
#>          series       sa    trend        seas        irr
#> Sep 2016 1393.5 1567.439 1554.386 -173.938673  13.052790
#> Oct 2016 1497.4 1547.052 1551.116  -49.651713  -4.064205
#> Nov 2016 1684.3 1544.888 1549.070  139.411575  -4.182012
#> Dec 2016 2850.4 1560.255 1548.078 1290.145096  12.177219
#> Jan 2017 1428.5 1529.584 1547.770 -101.084412 -18.185622
#> Feb 2017 1092.4 1552.751 1547.915 -460.350943   4.835817
#> Mar 2017 1370.3 1532.112 1548.227 -161.812246 -16.114386
#> Apr 2017 1522.6 1571.429 1548.386  -48.829077  23.042779
#> May 2017 1452.4 1554.351 1548.282 -101.951014   6.069336
#> Jun 2017 1557.2 1548.139 1547.682    9.061423   0.456243
#> Jul 2017 1445.5 1540.321 1546.422  -94.820887  -6.101117
#> Aug 2017 1303.1 1554.506 1544.648 -251.405622   9.857735
sp = set_x11(sp,
             henderson.filter = 13)
x13_fast(y, spec = sp)
#> RegARIMA
#> Log-transformation: no 
#> SARIMA model:  (3,1,1) (0,1,1)
#> 
#> Coefficients
#>           Estimate Std. Error  T-stat
#> phi(1)     0.16557    0.06461   2.563
#> phi(2)     0.08780    0.06435   1.364
#> phi(3)    -0.11287    0.06047  -1.867
#> theta(1)  -0.86076    0.03927 -21.918
#> btheta(1) -0.22918    0.05193  -4.413
#> 
#> Regression model:
#>                  Estimate Std. Error T-stat
#> td                 1.0015     0.8098  1.237
#> lp                29.6392    11.5469  2.567
#> AO (2010-01-01)   37.0476    34.4300  1.076
#> AO (2015-01-01)   27.9946    35.0234  0.799
#> AO (2000-06-01)  199.7536    34.5756  5.777
#> AO (2000-07-01) -194.6887    34.6240 -5.623
#> LS (2005-04-01)  -82.3062    17.4210 -4.725
#> LS (2015-07-01)   81.3334    18.0924  4.495
#> Number of observations:  425 
#> Number of effective observations:  412 
#> Number of parameters:  14 
#> 
#> Loglikelihood:  -2159.902 
#> Standard error of the regression (ML estimate):  45.632 
#> AIC:  4347.804 
#> AICC:  4348.862 
#> BIC:  4404.099 
#> 
#> 
#> Decomposition
#> Monitoring and Quality Assessment Statistics: 
#>     M stats
#> m1    0.036
#> m2    0.029
#> m3    1.786
#> m4    0.537
#> m5    1.528
#> m6    0.449
#> m7    0.167
#> m8    0.290
#> m9    0.144
#> m10   0.297
#> m11   0.285
#> q     0.533
#> qm2   0.604
#> 
#> Final filters: 
#> Seasonal filter:  
#> Trend filter: 13 terms Henderson moving average
#> 
#> Diagnostics
#> Relative contribution of the components to the stationary
#> portion of the variance in the original series,
#> after the removal of the long term trend (in %)
#> 
#>            Component
#>  cycle         6.227
#>  seasonal     94.535
#>  irregular     0.290
#>  calendar      0.020
#>  others        1.506
#>  total       102.578
#> 
#> Residual seasonality tests
#>                 P.value
#>  seas.ftest.i     0.924
#>  seas.ftest.sa    0.997
#>  seas.qstest.i    0.018
#>  seas.qstest.sa   0.117
#>  td.ftest.i       0.000
#>  td.ftest.sa      0.000
#> 
#> 
#> Final
#> Last values
#>          series       sa    trend        seas        irr
#> Sep 2016 1393.5 1567.255 1554.069 -173.754955  13.186079
#> Oct 2016 1497.4 1547.824 1551.331  -50.424459  -3.506850
#> Nov 2016 1684.3 1546.128 1548.057  138.172189  -1.929197
#> Dec 2016 2850.4 1557.769 1545.941 1292.630528  11.828053
#> Jan 2017 1428.5 1530.142 1545.777 -101.642006 -15.634829
#> Feb 2017 1092.4 1552.860 1546.573 -460.459760   6.287131
#> Mar 2017 1370.3 1532.029 1548.853 -161.728778 -16.824581
#> Apr 2017 1522.6 1571.073 1551.472  -48.473184  19.601512
#> May 2017 1452.4 1554.849 1552.698 -102.448692   2.151163
#> Jun 2017 1557.2 1550.052 1552.154    7.148413  -2.102681
#> Jul 2017 1445.5 1541.002 1549.653  -95.502435  -8.650144
#> Aug 2017 1303.1 1554.802 1545.625 -251.702019   9.176569